Experimental(?) commit which introduces a small information window to the UI. This is also a first step in introducing "bridge" structures to expose pure C++ classes through to QML (without converting all objects to QObjects). Did a one stop in converting all double to float as well while at it :)
Oh, objects can be selected by clickin in the image. This has worked for some time but never exposed in the UI :)

+++ raytracer/main/point.cpp
@@ -1,9 +1,9 @@
 #include "point.h"
 #include <QDebug>
 
-Point::Point(double x,
-             double y,
-             double z) :
+Point::Point(float x,
+             float y,
+             float z) :
     Tuple(x, y, z, 1)
 {
     // qDebug() << "Created Point. x:" << x << " y:" << y << " z:" << z;
